+++ src/test/java/roomescape/domain/member/domain/RoleTest.java
@@ -1,10 +1,9 @@
-package roomescape.domain.member;
+package roomescape.domain.member.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.member.domain.Role;
 
 class RoleTest {
