+++ src/test/java/roomescape/domain/view/AdminViewControllerTest.java
@@ -1,7 +1,6 @@
 package roomescape.domain.view;
 
 import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
